pyproject.toml
[tool.poetry]
name = "education"
version = "0.1.0"
description = ""
authors = ["Maxim <max2004@zel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
django-debug-toolbar-force = "^0.2"
coverage = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
